Convert implementations from float to double
All floats become double
All Vector2f become Vector2d, All BoundingBox2f become BoundingBox2d

+++ src/barnes_hut/body.cpp
@@ -4,7 +4,7 @@
 
 namespace bh {
 
-Body::Body(Vector2f position, float mass)
+Body::Body(Vector2d position, double mass)
     : m_position(std::move(position)), m_mass(mass) {}
 
 }  // namespace bh
