+++ src/barnes_hut/force.cpp
@@ -17,37 +17,37 @@ namespace bh {
 #define OMEGA 0.5f
 #endif
 
-Vector2f compute_gravitational_force(const Body& b1, const Body& b2) {
-  Vector2f direction_v = b1.m_position - b2.m_position;
-  float distance = direction_v.norm();
+Vector2d compute_gravitational_force(const Body& b1, const Body& b2) {
+  Vector2d direction_v = b1.m_position - b2.m_position;
+  double distance = direction_v.norm();
   if (distance == 0) {
     return {0, 0};
   }
-  float magnitude = (G * b1.m_mass * b2.m_mass) / std::pow(distance, 2.f);
-  float angle = std::atan2(direction_v.y(), direction_v.x());
-  float fx = std::cos(angle) * magnitude;
-  float fy = std::sin(angle) * magnitude;
+  double magnitude = (G * b1.m_mass * b2.m_mass) / std::pow(distance, 2.f);
+  double angle = std::atan2(direction_v.y(), direction_v.x());
+  double fx = std::cos(angle) * magnitude;
+  double fy = std::sin(angle) * magnitude;
   return {fx, fy};
 }
 
-Vector2f compute_approximate_net_force_on_body(const Node& node,
+Vector2d compute_approximate_net_force_on_body(const Node& node,
                                                const Body& body) {
-  const auto visit_empty = [](const Empty&) -> Vector2f { return {0, 0}; };
+  const auto visit_empty = [](const Empty&) -> Vector2d { return {0, 0}; };
 
-  const auto visit_body = [body](const Body& visited) -> Vector2f {
+  const auto visit_body = [body](const Body& visited) -> Vector2d {
     return compute_gravitational_force(visited, body);
   };
 
-  const auto visit_region = [&](const Subquadrants& subquadrants) -> Vector2f {
-    float distance = (body.m_position - node.center_of_mass()).norm();
+  const auto visit_region = [&](const Subquadrants& subquadrants) -> Vector2d {
+    double distance = (body.m_position - node.center_of_mass()).norm();
     if (node.length() / distance < OMEGA) {
       // Approximation
       return bh::compute_gravitational_force(
           {node.center_of_mass(), node.total_mass()}, body);
     } else {
       return std::accumulate(
-          subquadrants.begin(), subquadrants.end(), Vector2f{0, 0},
-          [body](const Vector2f& total, const std::unique_ptr<Node>& curr) {
+          subquadrants.begin(), subquadrants.end(), Vector2d{0, 0},
+          [body](const Vector2d& total, const std::unique_ptr<Node>& curr) {
             return (total + compute_approximate_net_force_on_body(*curr, body))
                 .eval();
           });
@@ -58,17 +58,17 @@ Vector2f compute_approximate_net_force_on_body(const Node& node,
                     node.data());
 }
 
-Vector2f compute_exact_net_force_on_body(const Node& node, const Body& body) {
-  const auto visit_empty = [](const Empty&) -> Vector2f { return {0, 0}; };
+Vector2d compute_exact_net_force_on_body(const Node& node, const Body& body) {
+  const auto visit_empty = [](const Empty&) -> Vector2d { return {0, 0}; };
 
-  const auto visit_body = [body](const Body& visited) -> Vector2f {
+  const auto visit_body = [body](const Body& visited) -> Vector2d {
     return compute_gravitational_force(visited, body);
   };
 
-  const auto visit_region = [&](const Subquadrants& subquadrants) -> Vector2f {
+  const auto visit_region = [&](const Subquadrants& subquadrants) -> Vector2d {
     return std::accumulate(
-        subquadrants.begin(), subquadrants.end(), Vector2f{0, 0},
-        [body](const Vector2f& total, const std::unique_ptr<Node>& curr) {
+        subquadrants.begin(), subquadrants.end(), Vector2d{0, 0},
+        [body](const Vector2d& total, const std::unique_ptr<Node>& curr) {
           return (total + compute_approximate_net_force_on_body(*curr, body))
               .eval();
         });
